+++ src/birdnet/__init__.py
@@ -1,4 +1,5 @@
 from birdnet.audio_based_prediction import predict_species_within_audio_file
+from birdnet.audio_based_prediction_mp import predict_species_within_audio_files_mp
 from birdnet.location_based_prediction import predict_species_at_location_and_time
 from birdnet.types import (Confidence, Language, Species, SpeciesPrediction, SpeciesPredictions,
                            TimeInterval)