+++ src/birdnet_tests/models/m2v4/model_v2m4_tflite_py/test_predict_species_within_audio_file.py
@@ -46,7 +46,7 @@ def test_invalid_species_filter_raises_value_error(model: AudioModelV2M4TFLite):
   with pytest.raises(ValueError, match=rf"At least one species defined in 'filter_species' is invalid! They need to be known species, e.g., {', '.join(model.species[:3])}"):
     next(predict_species_within_audio_file(
       TEST_FILE_WAV,
-      filter_species=invalid_filter_species,
+      species_filter=invalid_filter_species,
       custom_model=model,
     ))